refactor(integrations): extract github_callback package from integration views
Move the GitHub OAuth/install callback logic out of ``posthog.api.integration``
and ``posthog.api.user_integration`` into a new ``posthog.api.github_callback``
package, with one module per concern:

- ``types`` — shared constants (cache prefixes, mobile redirect URL,
  ``APP_CONNECT_FROM_VALUES``) and ``github_oauth_redirect_uri`` /
  ``is_valid_github_installation_id`` helpers.
- ``redirects`` — ``AppDeepLinkRedirect`` and ``final_github_redirect`` for the
  post-callback web vs ``posthog://`` mobile vs ``/account-connected`` desktop
  routing.
- ``personal_state`` — state-token plumbing and off-host URL builders for the
  personal ``UserIntegration`` linking flow.
- ``team_services`` — three primitives previously inlined in the team-level
  ``IntegrationViewSet``: creating an integration from an OAuth ``code``
  (``IntegrationSerializer.create``), reusing a sibling-team install
  (``github_link_existing``), and minting a fresh OAuth URL when the install
  flow returns without one (``github_oauth_authorize``).
- ``personal_finish`` — the ``github_link_complete`` view body, now imported by
  ``posthog/urls.py`` directly.

Behavior is unchanged: same cache key schemas (``github_state:{user_id}`` for
``IntegrationViewSet.authorize`` and ``github_user_install_state:{token}`` for
the personal flow), same error codes (``access_denied`` /
``github_oauth_error`` / ``invalid_state`` / ``invalid_installation`` /
``installation_not_authorized`` / ``installation_token_failed`` / …), and same
``connect_from`` routing. No new state schemas, dispatchers, or recovery
flows — those land in a follow-up PR.

Frontend, MCP definitions, and generated types are untouched.

Generated-By: PostHog Code
Task-Id: d630d159-265e-4bfd-8cbe-c75c0de19c15
