+++ posthog/api/github_callback/__init__.py
@@ -1,0 +1,1 @@
+"""GitHub OAuth/App callback handling, split out for clarity from ``posthog.api.integration``."""
