+++ posthog/api/user_integration.py
@@ -11,15 +11,10 @@ Login management is fully handled by ``UserSocialAuth`` (python-social-auth) and
 is not controlled here.
 """
 
-import re
 from typing import Any, cast
-from urllib.parse import parse_qs, urlencode
+from urllib.parse import urlencode
 
-from django.conf import settings
 from django.core.cache import cache
-from django.http import HttpRequest, HttpResponseRedirect
-from django.shortcuts import redirect
-from django.views.decorators.http import require_http_methods
 
 import requests
 import structlog
@@ -30,150 +25,32 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
 
-from posthog.api.integration import (
+from posthog.api.github_callback.personal_state import github_app_install_url, github_oauth_authorize_url
+from posthog.api.github_callback.types import (
+    APP_CONNECT_FROM_VALUES,
     GITHUB_INSTALL_STATE_CACHE_PREFIX,
     GITHUB_INSTALL_STATE_TTL_SECONDS,
+)
+from posthog.api.integration import (
     GitHubBranchesQuerySerializer,
     GitHubBranchesResponseSerializer,
     GitHubReposQuerySerializer,
     GitHubReposRefreshResponseSerializer,
     GitHubReposResponseSerializer,
-    github_oauth_redirect_uri,
     validate_github_repository_name,
 )
 from posthog.auth import (
     OAuthAccessTokenAuthentication,
     PersonalAPIKeyAuthentication,
     SessionAuthentication,
-    session_auth_required,
 )
-from posthog.models.instance_setting import get_instance_settings
-from posthog.models.integration import (
-    GITHUB_REPOSITORY_REFRESH_COOLDOWN_SECONDS,
-    GitHubInstallationAccess,
-    GitHubIntegration,
-    Integration,
-)
+from posthog.models.integration import GITHUB_REPOSITORY_REFRESH_COOLDOWN_SECONDS, Integration
 from posthog.models.user import User
-from posthog.models.user_integration import (
-    UserGitHubIntegration,
-    UserIntegration,
-    user_github_integration_from_installation,
-)
+from posthog.models.user_integration import UserGitHubIntegration, UserIntegration
 from posthog.permissions import APIScopePermission
 from posthog.rate_limit import UserAuthenticationThrottle
 
 logger = structlog.get_logger(__name__)
-
-PERSONAL_INTEGRATIONS_SETTINGS_PATH = "/settings/user-personal-integrations"
-ACCOUNT_CONNECTED_GITHUB_INTEGRATION_PATH = "/account-connected/github-integration"
-# Native deep link the mobile app (apps/mobile) registers as the OAuth return
-# URL. The in-app browser (ASWebAuthenticationSession / Custom Tabs) closes and
-# returns control to the app when the callback 302s here.
-MOBILE_GITHUB_CALLBACK_URL = "posthog://github/callback"
-# ``connect_from`` values for first-party clients that use the lightweight app
-# linking flow (OAuth-only when the team already has the GitHub App installed,
-# otherwise discover/install) and return to a client-specific destination.
-APP_CONNECT_FROM_VALUES = ("posthog_code", "posthog_mobile")
-
-
-class _AppDeepLinkRedirect(HttpResponseRedirect):
-    """Redirect that also permits the mobile app's custom ``posthog://`` scheme.
-
-    Django's default ``HttpResponseRedirect`` rejects non-web schemes as unsafe
-    (``DisallowedRedirect``). The target here is a hardcoded first-party deep
-    link, not user input, so allowing the extra scheme is safe.
-    """
-
-    allowed_schemes = [*HttpResponseRedirect.allowed_schemes, "posthog"]
-
-
-def _final_github_redirect(connect_from: str | None, *, error: str | None = None) -> HttpResponseRedirect:
-    """Pick the post-OAuth destination based on which client started the flow.
-
-    - ``posthog_mobile`` → the app's ``posthog://`` deep link so the in-app
-      browser auto-closes.
-    - ``posthog_code`` → the web ``/account-connected`` page that the desktop app
-      intercepts via its own deep link.
-    - anything else (web UI) → the personal integrations settings page.
-    """
-    app_base_urls: dict[str, str] = {
-        "posthog_mobile": MOBILE_GITHUB_CALLBACK_URL,
-        "posthog_code": ACCOUNT_CONNECTED_GITHUB_INTEGRATION_PATH,
-    }
-    if connect_from in app_base_urls:
-        params = {"provider": "github"}
-        if error:
-            params["error"] = error
-        return _AppDeepLinkRedirect(f"{app_base_urls[connect_from]}?{urlencode(params)}")
-    if error:
-        return redirect(f"{PERSONAL_INTEGRATIONS_SETTINGS_PATH}?github_link_error={error}")
-    return redirect(f"{PERSONAL_INTEGRATIONS_SETTINGS_PATH}?github_link_success=1")
-
-
-def _github_oauth_authorize_url(state: str) -> str:
-    """Build the GitHub App user authorization URL."""
-    if not settings.GITHUB_APP_CLIENT_ID:
-        raise exceptions.ValidationError("GitHub App client ID is not configured (GITHUB_APP_CLIENT_ID missing).")
-    return "https://github.com/login/oauth/authorize?" + urlencode(
-        {"client_id": settings.GITHUB_APP_CLIENT_ID, "redirect_uri": github_oauth_redirect_uri(), "state": state}
-    )
-
-
-def _github_app_install_url(state: str) -> str:
-    """Build the GitHub App install URL."""
-    instance_settings = get_instance_settings(["GITHUB_APP_SLUG"])
-    app_slug = instance_settings.get("GITHUB_APP_SLUG")
-    if not app_slug:
-        raise exceptions.ValidationError("GitHub App is not configured on this instance (missing GITHUB_APP_SLUG).")
-    return f"https://github.com/apps/{app_slug}/installations/new?{urlencode({'state': state})}"
-
-
-def _github_state_token(state_raw: str) -> str:
-    state_params = parse_qs(state_raw)
-    return state_params["token"][0] if "token" in state_params else state_raw
-
-
-def _github_user_installation_ids(user_access_token: str) -> list[str]:
-    """List GitHub App installations visible to a user-to-server token."""
-    response = requests.get(
-        "https://api.github.com/user/installations",
-        headers={
-            "Accept": "application/vnd.github+json",
-            "Authorization": f"Bearer {user_access_token}",
-            "X-GitHub-Api-Version": "2022-11-28",
-        },
-        params={"per_page": 100},
-        timeout=10,
-    )
-    if response.status_code != 200:
-        logger.warning("github_link: failed to list user installations", status_code=response.status_code)
-        raise requests.RequestException(f"Unexpected status {response.status_code} listing user installations")
-
-    installations = response.json().get("installations", [])
-    ids: list[str] = []
-    if isinstance(installations, list):
-        for installation in installations:
-            if isinstance(installation, dict) and installation.get("id") is not None:
-                ids.append(str(installation["id"]))
-    return ids
-
-
-def _redirect_to_github_app_install(user: User, connect_from: str | None) -> HttpResponseRedirect:
-    """Continue from user OAuth discovery to app installation when no installation exists yet."""
-    from django.utils.crypto import get_random_string
-
-    token = get_random_string(48)
-    state = urlencode({"token": token, "source": "user_integration"})
-    install_state_payload: dict[str, Any] = {"user_id": user.id}
-    if connect_from:
-        install_state_payload["connect_from"] = connect_from
-    cache.set(
-        f"{GITHUB_INSTALL_STATE_CACHE_PREFIX}{token}",
-        install_state_payload,
-        timeout=GITHUB_INSTALL_STATE_TTL_SECONDS,
-    )
-    return redirect(_github_app_install_url(state))
 
 
 class UserGitHubAccountSerializer(serializers.Serializer):
@@ -449,7 +326,7 @@ class UserIntegrationViewSet(viewsets.GenericViewSet):
                     {"user_id": user.id, "connect_from": connect_from, "flow": "oauth_discover"},
                     timeout=GITHUB_INSTALL_STATE_TTL_SECONDS,
                 )
-                return Response({"install_url": _github_oauth_authorize_url(state), "connect_flow": "oauth_discover"})
+                return Response({"install_url": github_oauth_authorize_url(state), "connect_flow": "oauth_discover"})
 
         # If the user already has linked integrations, check whether there are
         # any GitHub App installations they haven't linked yet. If everything
@@ -470,235 +347,11 @@ class UserIntegrationViewSet(viewsets.GenericViewSet):
         )
         return Response(
             {
-                "install_url": _github_app_install_url(state),
+                "install_url": github_app_install_url(state),
                 "connect_flow": "app_install",
             }
         )
 
-
-def _app_connect_from_from_state_query(request: HttpRequest) -> str | None:
-    """Best-effort: OAuth error callbacks may still include ``state``; read cache
-    (do not delete) to recover which first-party client started the flow so the
-    error redirect returns to the right place."""
-    state_raw = request.GET.get("state")
-    if not state_raw:
-        return None
-    token = _github_state_token(state_raw)
-    cache_key = f"{GITHUB_INSTALL_STATE_CACHE_PREFIX}{token}"
-    state_payload = cache.get(cache_key)
-    if not state_payload or state_payload.get("user_id") != request.user.id:
-        return None
-    connect_from = state_payload.get("connect_from")
-    return connect_from if connect_from in APP_CONNECT_FROM_VALUES else None
-
-
-@require_http_methods(["GET"])
-@session_auth_required
-def github_link_complete(request: HttpRequest) -> HttpResponseRedirect:
-    """GitHub App installation + user authorization callback.
-
-    **Install flow** (``/installations/new``): GitHub sends ``installation_id``,
-    ``code``, and ``state`` in the query string.
-
-    **OAuth-only flow** (``/login/oauth/authorize`` when the team already has the
-    app installed): GitHub sends only ``code`` and ``state``; ``installation_id``
-    is read from validated server-side state.
-
-    Steps: validate state, resolve ``installation_id``, exchange ``code`` for
-    user-to-server tokens, fetch installation token, create/update ``UserIntegration``.
-    """
-
-    user = cast(User, request.user)
-
-    # Which first-party client (if any) started this flow; controls the final
-    # redirect destination. Resolved from validated server-side state below.
-    connect_from_value: str | None = None
-
-    def _error(reason: str) -> HttpResponseRedirect:
-        logger.warning("github_link: redirecting with error", reason=reason, user_id=user.id)
-        return _final_github_redirect(connect_from_value, error=reason)
-
-    # GitHub appends ?error=... when the user denied consent.
-    if github_error := request.GET.get("error"):
-        logger.warning(
-            "github_link: GitHub returned error on callback",
-            error=github_error,
-            description=request.GET.get("error_description"),
-            user_id=user.id,
-        )
-        connect_from_value = _app_connect_from_from_state_query(request)
-        return _error(github_error if github_error == "access_denied" else "github_oauth_error")
-
-    code = request.GET.get("code")
-    state_raw = request.GET.get("state")
-
-    if not code or not state_raw:
-        return _error("missing_params")
-
-    # The frontend extracts the raw token from the URL-encoded state before forwarding
-    # here, so state_raw is normally the 48-char random token.  Handle both forms so
-    # direct backend calls (e.g. in tests) and any future flow changes work correctly.
-    token = _github_state_token(state_raw)
-
-    # Validate state
-    cache_key = f"{GITHUB_INSTALL_STATE_CACHE_PREFIX}{token}"
-    state_payload = cache.get(cache_key)
-    if not state_payload or state_payload.get("user_id") != user.id:
-        return _error("invalid_state")
-    connect_from_value = state_payload.get("connect_from")
-    cache.delete(cache_key)
-
-    flow = state_payload.get("flow")
-    oauth_flow = flow == "oauth_authorize"
-    oauth_discover_flow = flow == "oauth_discover"
-    team_oauth_flow = flow == "team_oauth_authorize"
-    team_oauth_team_id: int | None = None
-    team_oauth_next: str | None = None
-    installation_ids: list[str] = []
-    if oauth_flow:
-        installation_id = state_payload.get("installation_id")
-        if not installation_id or not isinstance(installation_id, str):
-            return _error("missing_params")
-        # Do not require ``current_team`` to match: OAuth completes in a browser
-        # session whose active team may differ from the app that started the flow.
-        if not Integration.objects.filter(
-            kind="github", integration_id=installation_id, team__in=user.teams.all()
-        ).exists():
-            return _error("invalid_installation")
-        installation_ids = [installation_id]
-    elif team_oauth_flow:
-        # Triggered when the GitHub App is already installed on the org but no PostHog
-        # team has captured it (orphan installation). The frontend redirects the user
-        # through GitHub's User OAuth so we get a `code` for verify_user_installation_access.
-        installation_id = state_payload.get("installation_id")
-        team_oauth_team_id = state_payload.get("team_id")
-        team_oauth_next = state_payload.get("next") or None
-        if not installation_id or not isinstance(installation_id, str):
-            return _error("missing_params")
-        if not isinstance(team_oauth_team_id, int):
-            return _error("invalid_state")
-        # Confirm the user still has access to the team they started this flow from.
-        if not user.teams.filter(id=team_oauth_team_id).exists():
-            return _error("invalid_team")
-        installation_ids = [installation_id]
-    elif oauth_discover_flow:
-        pass
-    else:
-        installation_id = request.GET.get("installation_id")
-        if not installation_id:
-            return _error("missing_params")
-        installation_ids = [installation_id]
-
-    # Exchange code for user-to-server tokens
-    if oauth_flow or oauth_discover_flow or team_oauth_flow:
-        authorization = GitHubIntegration.github_user_from_code(code, redirect_uri=github_oauth_redirect_uri())
-    else:
-        authorization = GitHubIntegration.github_user_from_code(code)
-    if authorization is None:
-        return _error("exchange_failed")
-
-    if oauth_discover_flow:
-        try:
-            installation_ids = _github_user_installation_ids(authorization.access_token)
-        except requests.RequestException:
-            return _error("installation_fetch_failed")
-        if not installation_ids:
-            return _redirect_to_github_app_install(user, connect_from_value)
-
-    for installation_id in installation_ids:
-        # installation_id must be a plain positive integer (GitHub App IDs always are).
-        # Reject anything else before it touches URL construction.
-        if not re.fullmatch(r"\d{1,20}", str(installation_id)):
-            return _error("invalid_installation_id")
-
-    for installation_id in installation_ids:
-        installation_id = str(installation_id)
-        # Verify the authorizing user actually has access to this installation.
-        # An attacker could use their own OAuth code with someone else's installation_id
-        # to obtain an installation token scoped to a different organisation's repos.
-        if not oauth_discover_flow:
-            try:
-                has_access = GitHubIntegration.verify_user_installation_access(
-                    installation_id, authorization.access_token
-                )
-            except requests.RequestException:
-                logger.warning(
-                    "github_link: installation ownership check failed",
-                    installation_id=installation_id,
-                    user_id=user.id,
-                    exc_info=True,
-                )
-                return _error("installation_verify_failed")
-            if not has_access:
-                logger.warning(
-                    "github_link: user does not have access to installation",
-                    installation_id=installation_id,
-                    user_id=user.id,
-                )
-                return _error("installation_not_authorized")
-
-        # Get installation info and access token
-        try:
-            installation_info = GitHubIntegration.client_request(f"installations/{installation_id}").json()
-            access_token_response = GitHubIntegration.client_request(
-                f"installations/{installation_id}/access_tokens", method="POST"
-            ).json()
-        except Exception:
-            logger.warning("github_link: failed to fetch installation info", exc_info=True)
-            return _error("installation_fetch_failed")
-
-        installation_access_token = access_token_response.get("token")
-        if not installation_access_token:
-            return _error("installation_token_failed")
-
-        token_expires_at = access_token_response.get("expires_at")
-        if not token_expires_at:
-            return _error("installation_token_failed")
-
-        user_github_integration_from_installation(
-            user,
-            GitHubInstallationAccess(
-                installation_id=installation_id,
-                installation_info=installation_info,
-                access_token=installation_access_token,
-                token_expires_at=token_expires_at,
-                repository_selection=access_token_response.get("repository_selection", "selected"),
-            ),
-            authorization,
-        )
-
-    if team_oauth_flow and team_oauth_team_id is not None:
-        installation_id = str(installation_ids[0])
-        # Create the team-level Integration that the user originally tried to install.
-        # ``integration_from_installation_id`` re-fetches the installation token, so it
-        # works whether or not another team in the org has already linked this install.
-        try:
-            team_integration = GitHubIntegration.integration_from_installation_id(
-                installation_id, team_oauth_team_id, user
-            )
-        except Exception:
-            logger.warning(
-                "github_link: failed to create team integration",
-                installation_id=installation_id,
-                team_id=team_oauth_team_id,
-                exc_info=True,
-            )
-            return _error("integration_create_failed")
-
-        # Mirror the fresh-install flow: stamp the connecting user's GitHub login on
-        # the team integration card.
-        team_integration.config["connecting_user_github_login"] = authorization.gh_login
-        team_integration.save(update_fields=["config"])
-
-        target = team_oauth_next or PERSONAL_INTEGRATIONS_SETTINGS_PATH
-        forwarded_params: dict[str, str] = {
-            "installation_id": installation_id,
-            "integration_id": str(team_integration.id),
-        }
-        joiner = "&" if "?" in target else "?"
-        return redirect(f"{target}{joiner}{urlencode(forwarded_params)}")
-
-    return _final_github_redirect(connect_from_value)
 
 
 def _resolve_team_for_github_start(user: User, request: Request):
@@ -811,7 +464,7 @@ def _attempt_app_oauth_fast_path(
         },
         timeout=GITHUB_INSTALL_STATE_TTL_SECONDS,
     )
-    return Response({"install_url": _github_oauth_authorize_url(state), "connect_flow": "oauth_authorize"})
+    return Response({"install_url": github_oauth_authorize_url(state), "connect_flow": "oauth_authorize"})
 
 
 def _serialize_github_integration(
